requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
Pillow>=10.0,<11.0
psycopg2-binary>=2.9,<3.0
